PP-5360: Clean up
* move Link back to where it originally was

* removed not needed properties from link

* removed unused methods from RefundSummary
