--- src/test/resources/pacts/publicapi-ledger-get-payment-with-awaiting-capture-request-state.json
@@ -1,109 +1,0 @@
-{
-  "consumer": {
-    "name": "publicapi"
-  },
-  "provider": {
-    "name": "ledger"
-  },
-  "interactions": [
-    {
-      "description": "get a charge request with awaiting charge request state",
-      "providerStates": [
-        {
-          "name": "a transaction with delayed capture true and awaiting capture request status exists",
-          "params": {
-            "charge_id": "ch_123abc456xyz"
-          }
-        }
-      ],
-      "request": {
-        "method": "GET",
-        "path": "/v1/transaction/ch_123abc456xyz"
-      },
-      "response": {
-        "status": 200,
-        "headers": {
-          "Content-Type": "application/json"
-        },
-        "body": {
-          "amount": 100,
-          "state": {
-            "finished": false,
-            "status": "submitted"
-          },
-          "description": "Test description",
-          "reference": "aReference",
-          "language": "en",
-          "links": [
-            {
-              "rel": "self",
-              "method": "GET",
-              "href": "https://ledger/v1/ledger/ch_123abc456xyz"
-            },
-            {
-              "rel": "refunds",
-              "method": "GET",
-              "href": "https://ledger/v1/transaction/ch_123abc456xyz/refunds"
-            }
-          ],
-          "charge_id": "ch_123abc456xyz",
-          "return_url": "https://somewhere.gov.uk/rainbow/1",
-          "payment_provider": "sandbox",
-          "created_date": "2018-10-16T10:46:02.121Z",
-          "refund_summary": {
-            "status": "unavailable",
-            "user_external_id": null,
-            "amount_available": 100,
-            "amount_submitted": 0
-          },
-          "settlement_summary": {
-            "capture_submit_time": null,
-            "captured_date": null
-          },
-          "delayed_capture": true
-        },
-        "matchingRules": {
-          "body": {
-            "$.reference": {
-              "matchers": [{"match": "type"}]
-            },
-            "$.description": {
-              "matchers": [{"match": "type"}]
-            },
-            "$.links[0].href": {
-              "matchers": [
-                {
-                  "regex": "http.*:\/\/.*\/v1\/transaction\/ch_123abc456xyz"
-                }
-              ]
-            },
-            "$.links[1].href": {
-              "matchers": [
-                {
-                  "regex": "http.*:\/\/.*\/v1\/transaction\/ch_123abc456xyz\/refunds"
-                }
-              ]
-            },
-            "$.return_url": {
-              "matchers": [{"match": "type"}]
-            },
-            "$.created_date": {
-              "matchers": [{ "date": "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'" }]
-            },
-            "$.refund_summary.amount_available": {
-              "matchers": [{"match": "type"}]
-            }
-          }
-        }
-      }
-    }
-  ],
-  "metadata": {
-    "pact-specification": {
-      "version": "3.0.0"
-    },
-    "pact-jvm": {
-      "version": "3.5.16"
-    }
-  }
-}
